Updated node neighbors
Makes it so it sets node neighbors to all nodes in a way adjacent to it, need to make it just 1 node.

+++ include/algorithms/Algorithm.h
@@ -7,8 +7,8 @@
 class Algorithm {
 
 public:
-	std::unordered_set<int> visited;
-	std::vector<int> toVisit;
+	std::unordered_set<long long> visited;
+	std::vector<long long> toVisit;
 
 	virtual ~Algorithm() {};
 };
